+++ staticFiles/js/portal/src/home.js
@@ -1,7 +1,8 @@
+"use strict";
+
 (function($){
     $(function(){
         let $save = $(".save");
-        let targetPage = "platform";
 
         const ajaxData = path => {
           $.ajax({
@@ -39,7 +40,6 @@
                 $currentCataMark.show();
                 $currentCataMark.siblings().hide();
             }
-            targetPage=$(this).attr("data-type");
             ajaxData($this.attr("data-path"));
         });
 
@@ -81,9 +81,8 @@
         // 点击产品名
         $(".content").on("click",'.productName',function(){
             // 跳转到TRU平台导航栏中对应的该产品的信息
-            // const id = $(this).attr("data-id");
-            // window.location.href = `/workflow/${id}`;
-            window.location.href = "/portal/" + targetPage;
+            const id = $(this).attr("data-id");
+            window.location.href = `/portal/workflow/${id}`;
         });
 
         // 点击产品链接